+++ winter/web/pagination/page_processor_resolver.py
@@ -1,4 +1,4 @@
-import typing
+from typing import Any
 
 from winter.data.pagination import Page
 from winter.web.output_processor import IOutputProcessorResolver
@@ -10,8 +10,8 @@ class PageOutputProcessorResolver(IOutputProcessorResolver):
     def __init__(self):
         self._page_processor = PageProcessor()
 
-    def is_supported(self, body: typing.Any) -> bool:
+    def is_supported(self, body: Any) -> bool:
         return isinstance(body, Page)
 
-    def get_processor(self, body: typing.Any) -> PageProcessor:
+    def get_processor(self, body: Any) -> PageProcessor:
         return self._page_processor